requirements/dev.txt
Django==3.2.9
psycopg2-binary==2.9.1
factory-boy==3.2.0
django-extensions==3.1.3
ipython==7.28.0
flake8==3.9.2
mypy==0.910
django-stubs==1.9.0
tblib==1.7.0
bandit==1.7.0
coverage==6.0.1
python-dateutil==2.8.2
freezegun==1.1.0
types-freezegun==1.1.0
types-python-dateutil==2.8.2
